refactor: 删除 security models/ Potemkin 层（#326，SecurityConfig carve-out）
补审 #313 confirmed：security/src/models/（~1085 行）零消费 Potemkin 层——
src/acs/ 与 src/security/（真实实现）从不 import models::*，74 处 execute()
返 Value，4 处 typed 返回用 API 文件本地类型。794 行契约测试给死类型盖
"CI 绿=活着"戳。唯一 live 类型 SecurityConfig。

清理：
- SecurityConfig carve-out 到新 src/config.rs（从 models/mod.rs:7-72 移）
- 删 src/models/ 整目录（acs.rs/common.rs/security_and_compliance.rs/mod.rs 死类型）
- 删 794 行契约测试 tests/security_contract_models.rs
- lib.rs：pub mod models → config；crate::models::SecurityConfig → crate::config::
  SecurityConfig（5 处）；prelude glob → 显式 SecurityConfig；doc 架构图更新
- serde/serde_json 仍被 error.rs + acs/security 用，不动 Cargo.toml

验证：build/clippy(--no-default-features)/test(-p security --all-features)/fmt 全绿，
零残留 grep。

Closes #326
